+++ plugin_base/src/plugin_base.vst3/plugin_base.vst3/pb_editor.cpp
@@ -34,7 +34,9 @@ tresult PLUGIN_API
 pb_editor::checkSizeConstraint(ViewRect* new_size)
 {
   auto const& topo = *_controller->gui_state().desc().plugin;
-  int new_width = std::max(new_size->getWidth(), topo.gui.min_width);
+  int min_width = (int)(topo.gui.default_width * topo.gui.min_scale);
+  int max_width = (int)(topo.gui.default_width * topo.gui.max_scale);
+  int new_width = std::clamp((int)new_size->getWidth(), min_width, max_width);
   new_size->right = new_size->left + new_width;
   new_size->bottom = new_size->top + (new_width * topo.gui.aspect_ratio_height / topo.gui.aspect_ratio_width);
   return kResultTrue;