+++ plugin_base/src/plugin_base/plugin_base/topo/plugin.cpp
@@ -38,9 +38,10 @@ plugin_topo::validate() const
   assert(version.minor >= 0);
   assert(version.patch >= 0);
   assert(gui.typeface_file_name.size());
+  assert(0 < gui.min_scale && gui.min_scale <= gui.max_scale);
   assert(audio_polyphony >= 0 && audio_polyphony < topo_max);
   assert(graph_polyphony >= 0 && graph_polyphony < topo_max);
-  assert(0 < gui.min_width && gui.min_width <= 16384);
+  assert(0 < gui.default_width && gui.default_width <= 16384);
   assert(0 < gui.aspect_ratio_width && gui.aspect_ratio_width <= 1000);
   assert(0 < gui.aspect_ratio_height && gui.aspect_ratio_height <= 1000);
   assert(0 < gui.module_sections.size() && gui.module_sections.size() <= modules.size());