+++ src/firefly_synth/firefly_synth/synth.cpp
@@ -327,7 +327,7 @@ synth_topo(bool is_fx)
   result->voice_mode_module = module_voice_in;
   result->voice_mode_param = voice_in_param_mode;
 
-  result->gui.min_width = 1280;
+  result->gui.default_width = 1280;
   result->gui.aspect_ratio_width = 118;
   if(is_fx)
   {
@@ -349,7 +349,7 @@ synth_topo(bool is_fx)
 
   result->gui.typeface_file_name = "Handel Gothic Regular.ttf"; 
   result->gui.dimension.column_sizes = { is_fx? 30: 26, is_fx? 63: 67, 17, 32, 32 };
-  int height = result->gui.min_width * result->gui.aspect_ratio_height / result->gui.aspect_ratio_width;
+  int height = result->gui.default_width * result->gui.aspect_ratio_height / result->gui.aspect_ratio_width;
   std::vector<gui_vertical_section_size> section_vsizes = { { true, 1 }, { true, 1 }, { true, 2 }, { true, 2 } };
   if (!is_fx) section_vsizes.insert(section_vsizes.end(), { { true, 2 }, { true, 1 }, { true, 2 }, { true, 2 }, { true, 2 } });
   result->gui.dimension.row_sizes = gui_vertical_distribution(height, result->gui.font_height, section_vsizes);